requirements.txt
django==3.1.3
djangorestframework==3.12.2
requests
#PARA ESTA APLICACION DECIDI USAR LA LIBRERIA djangorestframework POR LO PRACTICA Y BIEN DOCUMENTADA QUE ESTA

# admin_monet
# Inicio*2021